♻️ [Refactoring]: merge executor.rs into deployment.rs
- Inline execute() and deploy_* impls back into deployment.rs root since splitting impls without a trait gave no real isolation benefit
- Rename executor_test.rs to deployment_test.rs to match the merged module
